requirements.txt
werkzeug
flask
opentelemetry-sdk
opentelemetry-exporter-prometheus-wsgi
prometheus_client